remove user table
