+++ src/main/java/com/thee5176/ledger_command/config/ModelMapperConfig.java
@@ -6,8 +6,8 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import com.thee5176.ledger_command.record.application.dto.LedgerItemsEntryDTO;
 import com.thee5176.ledger_command.record.application.dto.LedgersEntryDTO;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.pojos.LedgerItems;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.pojos.Ledgers;
+import com.thee5176.ledger_command.record.domain.model.tables.pojos.LedgerItems;
+import com.thee5176.ledger_command.record.domain.model.tables.pojos.Ledgers;
 
 @Configuration
 public class ModelMapperConfig {
