+++ src/main/java/com/thee5176/ledger_command/record/application/dto/LedgerItemsEntryDTO.java
@@ -1,6 +1,6 @@
 package com.thee5176.ledger_command.record.application.dto;
 
-import com.thee5176.ledger_command.record.domain.model.accounting.enums.BalanceType;
+import com.thee5176.ledger_command.record.domain.model.enums.BalanceType;
 import io.swagger.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Data;
