+++ src/main/java/com/thee5176/ledger_command/record/application/mapper/LedgerItemsMapper.java
@@ -5,7 +5,7 @@ import java.util.stream.Collectors;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import com.thee5176.ledger_command.record.application.dto.LedgersEntryDTO;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.pojos.LedgerItems;
+import com.thee5176.ledger_command.record.domain.model.tables.pojos.LedgerItems;
 
 @Service
 public class LedgerItemsMapper {
