+++ src/main/java/com/thee5176/ledger_command/record/domain/model/Accounting.java
@@ -1,13 +1,12 @@
 /*
  * This file is generated by jOOQ.
  */
-package com.thee5176.ledger_command.record.domain.model.accounting;
+package com.thee5176.ledger_command.record.domain.model;
 
 
-import com.thee5176.ledger_command.record.domain.model.DefaultCatalog;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.CodeOfAccount;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.LedgerItems;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.Ledgers;
+import com.thee5176.ledger_command.record.domain.model.tables.CodeOfAccount;
+import com.thee5176.ledger_command.record.domain.model.tables.LedgerItems;
+import com.thee5176.ledger_command.record.domain.model.tables.Ledgers;
 
 import java.util.Arrays;
 import java.util.List;
