+++ src/main/java/com/thee5176/ledger_command/record/domain/model/Tables.java
@@ -1,12 +1,12 @@
 /*
  * This file is generated by jOOQ.
  */
-package com.thee5176.ledger_command.record.domain.model.accounting;
+package com.thee5176.ledger_command.record.domain.model;
 
 
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.CodeOfAccount;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.LedgerItems;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.Ledgers;
+import com.thee5176.ledger_command.record.domain.model.tables.CodeOfAccount;
+import com.thee5176.ledger_command.record.domain.model.tables.LedgerItems;
+import com.thee5176.ledger_command.record.domain.model.tables.Ledgers;
 
 
 /**
