--- src/main/java/com/thee5176/ledger_command/record/domain/model/credential/Credential.java
@@ -1,63 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.thee5176.ledger_command.record.domain.model.credential;
-
-
-import com.thee5176.ledger_command.record.domain.model.DefaultCatalog;
-import com.thee5176.ledger_command.record.domain.model.credential.tables.Authorities;
-import com.thee5176.ledger_command.record.domain.model.credential.tables.Users;
-
-import java.util.Arrays;
-import java.util.List;
-
-import org.jooq.Catalog;
-import org.jooq.Table;
-import org.jooq.impl.DSL;
-import org.jooq.impl.SchemaImpl;
-
-
-/**
- * This class is generated by jOOQ.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
-public class Credential extends SchemaImpl {
-
-    private static final long serialVersionUID = 1L;
-
-    /**
-     * The reference instance of <code>credential</code>
-     */
-    public static final Credential CREDENTIAL = new Credential();
-
-    /**
-     * The table <code>credential.authorities</code>.
-     */
-    public final Authorities AUTHORITIES = Authorities.AUTHORITIES;
-
-    /**
-     * The table <code>credential.users</code>.
-     */
-    public final Users USERS = Users.USERS;
-
-    /**
-     * No further instances allowed
-     */
-    private Credential() {
-        super(DSL.name("credential"), null, DSL.comment(""));
-    }
-
-
-    @Override
-    public Catalog getCatalog() {
-        return DefaultCatalog.DEFAULT_CATALOG;
-    }
-
-    @Override
-    public final List<Table<?>> getTables() {
-        return Arrays.asList(
-            Authorities.AUTHORITIES,
-            Users.USERS
-        );
-    }
-}
