--- src/main/java/com/thee5176/ledger_command/record/domain/model/credential/Indexes.java
@@ -1,26 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.thee5176.ledger_command.record.domain.model.credential;
-
-
-import com.thee5176.ledger_command.record.domain.model.credential.tables.Authorities;
-
-import org.jooq.Index;
-import org.jooq.OrderField;
-import org.jooq.impl.DSL;
-import org.jooq.impl.Internal;
-
-
-/**
- * A class modelling indexes of tables in credential.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
-public class Indexes {
-
-    // -------------------------------------------------------------------------
-    // INDEX definitions
-    // -------------------------------------------------------------------------
-
-    public static final Index IX_AUTH_USERNAME = Internal.createIndex(DSL.name("ix_auth_username"), Authorities.AUTHORITIES, new OrderField[] { Authorities.AUTHORITIES.USERNAME, Authorities.AUTHORITIES.AUTHORITY }, true);
-}
