--- src/main/java/com/thee5176/ledger_command/record/domain/model/credential/Tables.java
@@ -1,26 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.thee5176.ledger_command.record.domain.model.credential;
-
-
-import com.thee5176.ledger_command.record.domain.model.credential.tables.Authorities;
-import com.thee5176.ledger_command.record.domain.model.credential.tables.Users;
-
-
-/**
- * Convenience access to all tables in credential.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
-public class Tables {
-
-    /**
-     * The table <code>credential.authorities</code>.
-     */
-    public static final Authorities AUTHORITIES = Authorities.AUTHORITIES;
-
-    /**
-     * The table <code>credential.users</code>.
-     */
-    public static final Users USERS = Users.USERS;
-}
