--- src/main/java/com/thee5176/ledger_command/record/domain/model/credential/tables/Users.java
@@ -1,335 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.thee5176.ledger_command.record.domain.model.credential.tables;
-
-
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.Ledgers.LedgersPath;
-import com.thee5176.ledger_command.record.domain.model.credential.Credential;
-import com.thee5176.ledger_command.record.domain.model.credential.Keys;
-import com.thee5176.ledger_command.record.domain.model.credential.tables.Authorities.AuthoritiesPath;
-import com.thee5176.ledger_command.record.domain.model.credential.tables.records.UsersRecord;
-
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.List;
-
-import org.jooq.Condition;
-import org.jooq.Field;
-import org.jooq.ForeignKey;
-import org.jooq.InverseForeignKey;
-import org.jooq.Name;
-import org.jooq.Path;
-import org.jooq.PlainSQL;
-import org.jooq.QueryPart;
-import org.jooq.Record;
-import org.jooq.SQL;
-import org.jooq.Schema;
-import org.jooq.Select;
-import org.jooq.Stringly;
-import org.jooq.Table;
-import org.jooq.TableField;
-import org.jooq.TableOptions;
-import org.jooq.UniqueKey;
-import org.jooq.impl.DSL;
-import org.jooq.impl.SQLDataType;
-import org.jooq.impl.TableImpl;
-
-
-/**
- * This class is generated by jOOQ.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
-public class Users extends TableImpl<UsersRecord> {
-
-    private static final long serialVersionUID = 1L;
-
-    /**
-     * The reference instance of <code>credential.users</code>
-     */
-    public static final Users USERS = new Users();
-
-    /**
-     * The class holding records for this type
-     */
-    @Override
-    public Class<UsersRecord> getRecordType() {
-        return UsersRecord.class;
-    }
-
-    /**
-     * The column <code>credential.users.id</code>.
-     */
-    public final TableField<UsersRecord, String> ID = createField(DSL.name("id"), SQLDataType.VARCHAR.nullable(false), this, "");
-
-    /**
-     * The column <code>credential.users.username</code>.
-     */
-    public final TableField<UsersRecord, String> USERNAME = createField(DSL.name("username"), SQLDataType.VARCHAR(50).nullable(false), this, "");
-
-    /**
-     * The column <code>credential.users.password</code>.
-     */
-    public final TableField<UsersRecord, String> PASSWORD = createField(DSL.name("password"), SQLDataType.VARCHAR(255).nullable(false), this, "");
-
-    /**
-     * The column <code>credential.users.is_enabled</code>.
-     */
-    public final TableField<UsersRecord, Boolean> IS_ENABLED = createField(DSL.name("is_enabled"), SQLDataType.BOOLEAN, this, "");
-
-    /**
-     * The column <code>credential.users.is_account_non_expired</code>.
-     */
-    public final TableField<UsersRecord, Boolean> IS_ACCOUNT_NON_EXPIRED = createField(DSL.name("is_account_non_expired"), SQLDataType.BOOLEAN, this, "");
-
-    /**
-     * The column <code>credential.users.is_credentials_non_expired</code>.
-     */
-    public final TableField<UsersRecord, Boolean> IS_CREDENTIALS_NON_EXPIRED = createField(DSL.name("is_credentials_non_expired"), SQLDataType.BOOLEAN, this, "");
-
-    /**
-     * The column <code>credential.users.is_account_non_locked</code>.
-     */
-    public final TableField<UsersRecord, Boolean> IS_ACCOUNT_NON_LOCKED = createField(DSL.name("is_account_non_locked"), SQLDataType.BOOLEAN, this, "");
-
-    /**
-     * The column <code>credential.users.firstname</code>.
-     */
-    public final TableField<UsersRecord, String> FIRSTNAME = createField(DSL.name("firstname"), SQLDataType.VARCHAR(255).nullable(false), this, "");
-
-    /**
-     * The column <code>credential.users.lastname</code>.
-     */
-    public final TableField<UsersRecord, String> LASTNAME = createField(DSL.name("lastname"), SQLDataType.VARCHAR(255).nullable(false), this, "");
-
-    /**
-     * The column <code>credential.users.email</code>.
-     */
-    public final TableField<UsersRecord, String> EMAIL = createField(DSL.name("email"), SQLDataType.VARCHAR(255).nullable(false), this, "");
-
-    private Users(Name alias, Table<UsersRecord> aliased) {
-        this(alias, aliased, (Field<?>[]) null, null);
-    }
-
-    private Users(Name alias, Table<UsersRecord> aliased, Field<?>[] parameters, Condition where) {
-        super(alias, null, aliased, parameters, DSL.comment(""), TableOptions.table(), where);
-    }
-
-    /**
-     * Create an aliased <code>credential.users</code> table reference
-     */
-    public Users(String alias) {
-        this(DSL.name(alias), USERS);
-    }
-
-    /**
-     * Create an aliased <code>credential.users</code> table reference
-     */
-    public Users(Name alias) {
-        this(alias, USERS);
-    }
-
-    /**
-     * Create a <code>credential.users</code> table reference
-     */
-    public Users() {
-        this(DSL.name("users"), null);
-    }
-
-    public <O extends Record> Users(Table<O> path, ForeignKey<O, UsersRecord> childPath, InverseForeignKey<O, UsersRecord> parentPath) {
-        super(path, childPath, parentPath, USERS);
-    }
-
-    /**
-     * A subtype implementing {@link Path} for simplified path-based joins.
-     */
-    public static class UsersPath extends Users implements Path<UsersRecord> {
-
-        private static final long serialVersionUID = 1L;
-        public <O extends Record> UsersPath(Table<O> path, ForeignKey<O, UsersRecord> childPath, InverseForeignKey<O, UsersRecord> parentPath) {
-            super(path, childPath, parentPath);
-        }
-        private UsersPath(Name alias, Table<UsersRecord> aliased) {
-            super(alias, aliased);
-        }
-
-        @Override
-        public UsersPath as(String alias) {
-            return new UsersPath(DSL.name(alias), this);
-        }
-
-        @Override
-        public UsersPath as(Name alias) {
-            return new UsersPath(alias, this);
-        }
-
-        @Override
-        public UsersPath as(Table<?> alias) {
-            return new UsersPath(alias.getQualifiedName(), this);
-        }
-    }
-
-    @Override
-    public Schema getSchema() {
-        return aliased() ? null : Credential.CREDENTIAL;
-    }
-
-    @Override
-    public UniqueKey<UsersRecord> getPrimaryKey() {
-        return Keys.USERS_PKEY;
-    }
-
-    @Override
-    public List<UniqueKey<UsersRecord>> getUniqueKeys() {
-        return Arrays.asList(Keys.USERS_USERNAME_KEY);
-    }
-
-    private transient LedgersPath _ledgers;
-
-    /**
-     * Get the implicit to-many join path to the <code>accounting.ledgers</code>
-     * table
-     */
-    public LedgersPath ledgers() {
-        if (_ledgers == null)
-            _ledgers = new LedgersPath(this, null, com.thee5176.ledger_command.record.domain.model.accounting.Keys.LEDGERS__LEDGER_ITEMS_USER_FKEY.getInverseKey());
-
-        return _ledgers;
-    }
-
-    private transient AuthoritiesPath _authorities;
-
-    /**
-     * Get the implicit to-many join path to the
-     * <code>credential.authorities</code> table
-     */
-    public AuthoritiesPath authorities() {
-        if (_authorities == null)
-            _authorities = new AuthoritiesPath(this, null, Keys.AUTHORITIES__FK_AUTHORITIES_USERS.getInverseKey());
-
-        return _authorities;
-    }
-
-    @Override
-    public Users as(String alias) {
-        return new Users(DSL.name(alias), this);
-    }
-
-    @Override
-    public Users as(Name alias) {
-        return new Users(alias, this);
-    }
-
-    @Override
-    public Users as(Table<?> alias) {
-        return new Users(alias.getQualifiedName(), this);
-    }
-
-    /**
-     * Rename this table
-     */
-    @Override
-    public Users rename(String name) {
-        return new Users(DSL.name(name), null);
-    }
-
-    /**
-     * Rename this table
-     */
-    @Override
-    public Users rename(Name name) {
-        return new Users(name, null);
-    }
-
-    /**
-     * Rename this table
-     */
-    @Override
-    public Users rename(Table<?> name) {
-        return new Users(name.getQualifiedName(), null);
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    public Users where(Condition condition) {
-        return new Users(getQualifiedName(), aliased() ? this : null, null, condition);
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    public Users where(Collection<? extends Condition> conditions) {
-        return where(DSL.and(conditions));
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    public Users where(Condition... conditions) {
-        return where(DSL.and(conditions));
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    public Users where(Field<Boolean> condition) {
-        return where(DSL.condition(condition));
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    @PlainSQL
-    public Users where(SQL condition) {
-        return where(DSL.condition(condition));
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    @PlainSQL
-    public Users where(@Stringly.SQL String condition) {
-        return where(DSL.condition(condition));
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    @PlainSQL
-    public Users where(@Stringly.SQL String condition, Object... binds) {
-        return where(DSL.condition(condition, binds));
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    @PlainSQL
-    public Users where(@Stringly.SQL String condition, QueryPart... parts) {
-        return where(DSL.condition(condition, parts));
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    public Users whereExists(Select<?> select) {
-        return where(DSL.exists(select));
-    }
-
-    /**
-     * Create an inline derived table from this table
-     */
-    @Override
-    public Users whereNotExists(Select<?> select) {
-        return where(DSL.notExists(select));
-    }
-}
