--- src/main/java/com/thee5176/ledger_command/record/domain/model/credential/tables/pojos/Authorities.java
@@ -1,137 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package com.thee5176.ledger_command.record.domain.model.credential.tables.pojos;
-
-
-import java.io.Serializable;
-import java.util.Arrays;
-
-
-/**
- * This class is generated by jOOQ.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
-public class Authorities implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
-    private Long id;
-    private String username;
-    private String[] authority;
-
-    public Authorities() {}
-
-    public Authorities(Authorities value) {
-        this.id = value.id;
-        this.username = value.username;
-        this.authority = value.authority;
-    }
-
-    public Authorities(
-        Long id,
-        String username,
-        String[] authority
-    ) {
-        this.id = id;
-        this.username = username;
-        this.authority = authority;
-    }
-
-    /**
-     * Getter for <code>credential.authorities.id</code>.
-     */
-    public Long getId() {
-        return this.id;
-    }
-
-    /**
-     * Setter for <code>credential.authorities.id</code>.
-     */
-    public Authorities setId(Long id) {
-        this.id = id;
-        return this;
-    }
-
-    /**
-     * Getter for <code>credential.authorities.username</code>.
-     */
-    public String getUsername() {
-        return this.username;
-    }
-
-    /**
-     * Setter for <code>credential.authorities.username</code>.
-     */
-    public Authorities setUsername(String username) {
-        this.username = username;
-        return this;
-    }
-
-    /**
-     * Getter for <code>credential.authorities.authority</code>.
-     */
-    public String[] getAuthority() {
-        return this.authority;
-    }
-
-    /**
-     * Setter for <code>credential.authorities.authority</code>.
-     */
-    public Authorities setAuthority(String[] authority) {
-        this.authority = authority;
-        return this;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj)
-            return true;
-        if (obj == null)
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        final Authorities other = (Authorities) obj;
-        if (this.id == null) {
-            if (other.id != null)
-                return false;
-        }
-        else if (!this.id.equals(other.id))
-            return false;
-        if (this.username == null) {
-            if (other.username != null)
-                return false;
-        }
-        else if (!this.username.equals(other.username))
-            return false;
-        if (this.authority == null) {
-            if (other.authority != null)
-                return false;
-        }
-        else if (!Arrays.deepEquals(this.authority, other.authority))
-            return false;
-        return true;
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((this.id == null) ? 0 : this.id.hashCode());
-        result = prime * result + ((this.username == null) ? 0 : this.username.hashCode());
-        result = prime * result + ((this.authority == null) ? 0 : Arrays.deepHashCode(this.authority));
-        return result;
-    }
-
-    @Override
-    public String toString() {
-        StringBuilder sb = new StringBuilder("Authorities (");
-
-        sb.append(id);
-        sb.append(", ").append(username);
-        sb.append(", ").append(Arrays.deepToString(authority));
-
-        sb.append(")");
-        return sb.toString();
-    }
-}
