+++ src/main/java/com/thee5176/ledger_command/record/domain/model/enums/BalanceType.java
@@ -1,10 +1,10 @@
 /*
  * This file is generated by jOOQ.
  */
-package com.thee5176.ledger_command.record.domain.model.accounting.enums;
+package com.thee5176.ledger_command.record.domain.model.enums;
 
 
-import com.thee5176.ledger_command.record.domain.model.accounting.Accounting;
+import com.thee5176.ledger_command.record.domain.model.Accounting;
 
 import org.jooq.Catalog;
 import org.jooq.EnumType;
@@ -15,23 +15,15 @@ import org.jooq.Schema;
  * This class is generated by jOOQ.
  */
 @SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
-public enum Element implements EnumType {
+public enum BalanceType implements EnumType {
 
-    Assets("Assets"),
+    Debit("Debit"),
 
-    Liabilities("Liabilities"),
-
-    Equity("Equity"),
-
-    Revenue("Revenue"),
-
-    Expenses("Expenses"),
-
-    Other("Other");
+    Credit("Credit");
 
     private final String literal;
 
-    private Element(String literal) {
+    private BalanceType(String literal) {
         this.literal = literal;
     }
 
@@ -47,7 +39,7 @@ public enum Element implements EnumType {
 
     @Override
     public String getName() {
-        return "element";
+        return "balance_type";
     }
 
     @Override
@@ -60,7 +52,7 @@ public enum Element implements EnumType {
      * <code>null</code>, if no such value could be found, see {@link
      * EnumType#lookupLiteral(Class, String)}.
      */
-    public static Element lookupLiteral(String literal) {
-        return EnumType.lookupLiteral(Element.class, literal);
+    public static BalanceType lookupLiteral(String literal) {
+        return EnumType.lookupLiteral(BalanceType.class, literal);
     }
 }
