+++ src/main/java/com/thee5176/ledger_command/record/domain/model/enums/Element.java
@@ -1,10 +1,10 @@
 /*
  * This file is generated by jOOQ.
  */
-package com.thee5176.ledger_command.record.domain.model.accounting.enums;
+package com.thee5176.ledger_command.record.domain.model.enums;
 
 
-import com.thee5176.ledger_command.record.domain.model.accounting.Accounting;
+import com.thee5176.ledger_command.record.domain.model.Accounting;
 
 import org.jooq.Catalog;
 import org.jooq.EnumType;
@@ -15,15 +15,23 @@ import org.jooq.Schema;
  * This class is generated by jOOQ.
  */
 @SuppressWarnings({ "all", "unchecked", "rawtypes", "this-escape" })
-public enum BalanceType implements EnumType {
+public enum Element implements EnumType {
 
-    Debit("Debit"),
+    Assets("Assets"),
 
-    Credit("Credit");
+    Liabilities("Liabilities"),
+
+    Equity("Equity"),
+
+    Revenue("Revenue"),
+
+    Expenses("Expenses"),
+
+    Other("Other");
 
     private final String literal;
 
-    private BalanceType(String literal) {
+    private Element(String literal) {
         this.literal = literal;
     }
 
@@ -39,7 +47,7 @@ public enum BalanceType implements EnumType {
 
     @Override
     public String getName() {
-        return "balance_type";
+        return "element";
     }
 
     @Override
@@ -52,7 +60,7 @@ public enum BalanceType implements EnumType {
      * <code>null</code>, if no such value could be found, see {@link
      * EnumType#lookupLiteral(Class, String)}.
      */
-    public static BalanceType lookupLiteral(String literal) {
-        return EnumType.lookupLiteral(BalanceType.class, literal);
+    public static Element lookupLiteral(String literal) {
+        return EnumType.lookupLiteral(Element.class, literal);
     }
 }
