+++ src/main/java/com/thee5176/ledger_command/record/domain/model/tables/CodeOfAccount.java
@@ -1,15 +1,15 @@
 /*
  * This file is generated by jOOQ.
  */
-package com.thee5176.ledger_command.record.domain.model.accounting.tables;
+package com.thee5176.ledger_command.record.domain.model.tables;
 
 
-import com.thee5176.ledger_command.record.domain.model.accounting.Accounting;
-import com.thee5176.ledger_command.record.domain.model.accounting.Keys;
-import com.thee5176.ledger_command.record.domain.model.accounting.enums.BalanceType;
-import com.thee5176.ledger_command.record.domain.model.accounting.enums.Element;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.LedgerItems.LedgerItemsPath;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.records.CodeOfAccountRecord;
+import com.thee5176.ledger_command.record.domain.model.Accounting;
+import com.thee5176.ledger_command.record.domain.model.Keys;
+import com.thee5176.ledger_command.record.domain.model.enums.BalanceType;
+import com.thee5176.ledger_command.record.domain.model.enums.Element;
+import com.thee5176.ledger_command.record.domain.model.tables.LedgerItems.LedgerItemsPath;
+import com.thee5176.ledger_command.record.domain.model.tables.records.CodeOfAccountRecord;
 
 import java.util.Arrays;
 import java.util.Collection;
