+++ src/main/java/com/thee5176/ledger_command/record/domain/model/tables/LedgerItems.java
@@ -1,15 +1,15 @@
 /*
  * This file is generated by jOOQ.
  */
-package com.thee5176.ledger_command.record.domain.model.accounting.tables;
+package com.thee5176.ledger_command.record.domain.model.tables;
 
 
-import com.thee5176.ledger_command.record.domain.model.accounting.Accounting;
-import com.thee5176.ledger_command.record.domain.model.accounting.Keys;
-import com.thee5176.ledger_command.record.domain.model.accounting.enums.BalanceType;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.CodeOfAccount.CodeOfAccountPath;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.Ledgers.LedgersPath;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.records.LedgerItemsRecord;
+import com.thee5176.ledger_command.record.domain.model.Accounting;
+import com.thee5176.ledger_command.record.domain.model.Keys;
+import com.thee5176.ledger_command.record.domain.model.enums.BalanceType;
+import com.thee5176.ledger_command.record.domain.model.tables.CodeOfAccount.CodeOfAccountPath;
+import com.thee5176.ledger_command.record.domain.model.tables.Ledgers.LedgersPath;
+import com.thee5176.ledger_command.record.domain.model.tables.records.LedgerItemsRecord;
 
 import java.time.LocalDateTime;
 import java.util.Arrays;
