+++ src/main/java/com/thee5176/ledger_command/record/domain/model/tables/Ledgers.java
@@ -1,20 +1,17 @@
 /*
  * This file is generated by jOOQ.
  */
-package com.thee5176.ledger_command.record.domain.model.accounting.tables;
+package com.thee5176.ledger_command.record.domain.model.tables;
 
 
-import com.thee5176.ledger_command.record.domain.model.accounting.Accounting;
-import com.thee5176.ledger_command.record.domain.model.accounting.Keys;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.LedgerItems.LedgerItemsPath;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.records.LedgersRecord;
-import com.thee5176.ledger_command.record.domain.model.credential.tables.Users.UsersPath;
+import com.thee5176.ledger_command.record.domain.model.Accounting;
+import com.thee5176.ledger_command.record.domain.model.Keys;
+import com.thee5176.ledger_command.record.domain.model.tables.LedgerItems.LedgerItemsPath;
+import com.thee5176.ledger_command.record.domain.model.tables.records.LedgersRecord;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.Arrays;
 import java.util.Collection;
-import java.util.List;
 import java.util.UUID;
 
 import org.jooq.Condition;
@@ -160,23 +157,6 @@ public class Ledgers extends TableImpl<LedgersRecord> {
     @Override
     public UniqueKey<LedgersRecord> getPrimaryKey() {
         return Keys.TRANSACTIONS_PKEY;
-    }
-
-    @Override
-    public List<ForeignKey<LedgersRecord, ?>> getReferences() {
-        return Arrays.asList(Keys.LEDGERS__LEDGER_ITEMS_USER_FKEY);
-    }
-
-    private transient UsersPath _users;
-
-    /**
-     * Get the implicit join path to the <code>credential.users</code> table.
-     */
-    public UsersPath users() {
-        if (_users == null)
-            _users = new UsersPath(this, Keys.LEDGERS__LEDGER_ITEMS_USER_FKEY, null);
-
-        return _users;
     }
 
     private transient LedgerItemsPath _ledgerItems;
