+++ src/main/java/com/thee5176/ledger_command/record/domain/model/tables/pojos/CodeOfAccount.java
@@ -1,11 +1,11 @@
 /*
  * This file is generated by jOOQ.
  */
-package com.thee5176.ledger_command.record.domain.model.accounting.tables.pojos;
+package com.thee5176.ledger_command.record.domain.model.tables.pojos;
 
 
-import com.thee5176.ledger_command.record.domain.model.accounting.enums.BalanceType;
-import com.thee5176.ledger_command.record.domain.model.accounting.enums.Element;
+import com.thee5176.ledger_command.record.domain.model.enums.BalanceType;
+import com.thee5176.ledger_command.record.domain.model.enums.Element;
 
 import java.io.Serializable;
