+++ src/main/java/com/thee5176/ledger_command/record/domain/model/tables/pojos/LedgerItems.java
@@ -1,10 +1,10 @@
 /*
  * This file is generated by jOOQ.
  */
-package com.thee5176.ledger_command.record.domain.model.accounting.tables.pojos;
+package com.thee5176.ledger_command.record.domain.model.tables.pojos;
 
 
-import com.thee5176.ledger_command.record.domain.model.accounting.enums.BalanceType;
+import com.thee5176.ledger_command.record.domain.model.enums.BalanceType;
 
 import java.io.Serializable;
 import java.time.LocalDateTime;
