+++ src/main/java/com/thee5176/ledger_command/record/domain/model/tables/pojos/Ledgers.java
@@ -1,7 +1,7 @@
 /*
  * This file is generated by jOOQ.
  */
-package com.thee5176.ledger_command.record.domain.model.accounting.tables.pojos;
+package com.thee5176.ledger_command.record.domain.model.tables.pojos;
 
 
 import java.io.Serializable;
