+++ src/main/java/com/thee5176/ledger_command/record/domain/model/tables/records/CodeOfAccountRecord.java
@@ -1,12 +1,12 @@
 /*
  * This file is generated by jOOQ.
  */
-package com.thee5176.ledger_command.record.domain.model.accounting.tables.records;
+package com.thee5176.ledger_command.record.domain.model.tables.records;
 
 
-import com.thee5176.ledger_command.record.domain.model.accounting.enums.BalanceType;
-import com.thee5176.ledger_command.record.domain.model.accounting.enums.Element;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.CodeOfAccount;
+import com.thee5176.ledger_command.record.domain.model.enums.BalanceType;
+import com.thee5176.ledger_command.record.domain.model.enums.Element;
+import com.thee5176.ledger_command.record.domain.model.tables.CodeOfAccount;
 
 import org.jooq.Record1;
 import org.jooq.impl.UpdatableRecordImpl;
@@ -132,7 +132,7 @@ public class CodeOfAccountRecord extends UpdatableRecordImpl<CodeOfAccountRecord
     /**
      * Create a detached, initialised CodeOfAccountRecord
      */
-    public CodeOfAccountRecord(com.thee5176.ledger_command.record.domain.model.accounting.tables.pojos.CodeOfAccount value) {
+    public CodeOfAccountRecord(com.thee5176.ledger_command.record.domain.model.tables.pojos.CodeOfAccount value) {
         super(CodeOfAccount.CODE_OF_ACCOUNT);
 
         if (value != null) {
