+++ src/main/java/com/thee5176/ledger_command/record/domain/model/tables/records/LedgerItemsRecord.java
@@ -1,11 +1,11 @@
 /*
  * This file is generated by jOOQ.
  */
-package com.thee5176.ledger_command.record.domain.model.accounting.tables.records;
+package com.thee5176.ledger_command.record.domain.model.tables.records;
 
 
-import com.thee5176.ledger_command.record.domain.model.accounting.enums.BalanceType;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.LedgerItems;
+import com.thee5176.ledger_command.record.domain.model.enums.BalanceType;
+import com.thee5176.ledger_command.record.domain.model.tables.LedgerItems;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
@@ -166,7 +166,7 @@ public class LedgerItemsRecord extends UpdatableRecordImpl<LedgerItemsRecord> {
     /**
      * Create a detached, initialised LedgerItemsRecord
      */
-    public LedgerItemsRecord(com.thee5176.ledger_command.record.domain.model.accounting.tables.pojos.LedgerItems value) {
+    public LedgerItemsRecord(com.thee5176.ledger_command.record.domain.model.tables.pojos.LedgerItems value) {
         super(LedgerItems.LEDGER_ITEMS);
 
         if (value != null) {
