+++ src/main/java/com/thee5176/ledger_command/record/domain/model/tables/records/LedgersRecord.java
@@ -1,10 +1,10 @@
 /*
  * This file is generated by jOOQ.
  */
-package com.thee5176.ledger_command.record.domain.model.accounting.tables.records;
+package com.thee5176.ledger_command.record.domain.model.tables.records;
 
 
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.Ledgers;
+import com.thee5176.ledger_command.record.domain.model.tables.Ledgers;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -150,7 +150,7 @@ public class LedgersRecord extends UpdatableRecordImpl<LedgersRecord> {
     /**
      * Create a detached, initialised LedgersRecord
      */
-    public LedgersRecord(com.thee5176.ledger_command.record.domain.model.accounting.tables.pojos.Ledgers value) {
+    public LedgersRecord(com.thee5176.ledger_command.record.domain.model.tables.pojos.Ledgers value) {
         super(Ledgers.LEDGERS);
 
         if (value != null) {
