+++ src/main/java/com/thee5176/ledger_command/record/domain/service/LedgerCommandService.java
@@ -11,8 +11,8 @@ import org.springframework.transaction.annotation.Transactional;
 import com.thee5176.ledger_command.record.application.dto.LedgersEntryDTO;
 import com.thee5176.ledger_command.record.application.mapper.LedgerItemsMapper;
 import com.thee5176.ledger_command.record.application.mapper.LedgerMapper;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.pojos.LedgerItems;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.pojos.Ledgers;
+import com.thee5176.ledger_command.record.domain.model.tables.pojos.LedgerItems;
+import com.thee5176.ledger_command.record.domain.model.tables.pojos.Ledgers;
 import com.thee5176.ledger_command.record.infrastructure.repository.LedgerItemsRepository;
 import com.thee5176.ledger_command.record.infrastructure.repository.LedgerRepository;
 import jakarta.validation.constraints.NotNull;
