+++ src/test/java/com/thee5176/ledger_command/config/ModelMapperConfigTest.java
@@ -10,8 +10,8 @@ import com.thee5176.ledger_command.record.application.dto.LedgerItemsEntryDTO;
 import com.thee5176.ledger_command.record.application.dto.LedgerItemsEntryDTOTest;
 import com.thee5176.ledger_command.record.application.dto.LedgersEntryDTO;
 import com.thee5176.ledger_command.record.application.dto.LedgersEntryDTOTest;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.pojos.LedgerItems;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.pojos.Ledgers;
+import com.thee5176.ledger_command.record.domain.model.tables.pojos.LedgerItems;
+import com.thee5176.ledger_command.record.domain.model.tables.pojos.Ledgers;
 
 @SpringBootTest
 class ModelMapperConfigTest {
