+++ src/test/java/com/thee5176/ledger_command/record/application/mapper/LedgerItemsMapperTest.java
@@ -11,7 +11,7 @@ import org.modelmapper.ModelMapper;
 import com.thee5176.ledger_command.record.application.dto.LedgerItemsEntryDTO;
 import com.thee5176.ledger_command.record.application.dto.LedgersEntryDTO;
 import com.thee5176.ledger_command.record.application.dto.LedgersEntryDTOTest;
-import com.thee5176.ledger_command.record.domain.model.accounting.tables.pojos.LedgerItems;
+import com.thee5176.ledger_command.record.domain.model.tables.pojos.LedgerItems;
 
 class LedgerItemsMapperTest {
